large scale restructure, part 1...
