+++ src/pymorize/cli.py
@@ -11,14 +11,13 @@ from dask.distributed import Client
 from rich.traceback import install as rich_traceback_install
 from streamlit.web import cli as stcli
 
-from pymorize.fesom_1p4.nodes_to_levels import convert
-
 from . import _version, caching, dev_utils
-from .cmorizer import CMORizer
-from .filecache import fc
-from .logging import add_report_logger, logger
-from .ssh_tunnel import ssh_tunnel_cli
-from .validate import GENERAL_VALIDATOR, PIPELINES_VALIDATOR, RULES_VALIDATOR
+from .core.cmorizer import CMORizer
+from .core.filecache import fc
+from .core.logging import add_report_logger, logger
+from .core.ssh_tunnel import ssh_tunnel_cli
+from .core.validate import GENERAL_VALIDATOR, PIPELINES_VALIDATOR, RULES_VALIDATOR
+from .fesom_1p4.nodes_to_levels import convert
 
 MAX_FRAMES = int(os.environ.get("PYMORIZE_ERROR_MAX_FRAMES", 3))
 """
