+++ src/pymorize/core/aux_files.py
