+++ src/pymorize/core/pipeline.py
