+++ src/pymorize/core/rule.py
