+++ src/pymorize/core/utils.py
