+++ src/pymorize/std_lib/generic.py
