+++ src/pymorize/std_lib/global_attributes.py
