+++ src/pymorize/std_lib/timeaverage.py
