+++ src/pymorize/std_lib/variable_attributes.py
